行を ChangeMySocialsFormRow に分離し index を prop で受け取る形に変更
これにより handleBlur が自身の index を閉じ込められ、event からの対象行特定が不要になる。
