+++ client/component/form/change-my-socials-form/change-my-socials-form-row.scss
@@ -1,0 +1,17 @@
+@charset "utf-8";
+
+
+.root {
+  column-gap: 2zu;
+  display: flex;
+  align-items: center;
+}
+
+.type {
+  flex: 0 0 auto;
+  inline-size: 40zu;
+}
+
+.url {
+  flex: 1 1 auto;
+}
