+++ client/component/form/change-my-socials-form/change-my-socials-form.scss
@@ -12,18 +12,3 @@
   display: flex;
   flex-direction: column;
 }
-
-.row {
-  column-gap: 2zu;
-  display: flex;
-  align-items: center;
-}
-
-.type {
-  flex: 0 0 auto;
-  inline-size: 40zu;
-}
-
-.url {
-  flex: 1 1 auto;
-}
