+++ client/component/form/change-my-socials-form/change-my-socials-form.tsx
@@ -1,13 +1,12 @@
 /* eslint-disable react/jsx-closing-bracket-location */
 
 import {faCheck} from "@fortawesome/sharp-regular-svg-icons";
-import {FocusEvent, ReactElement, useCallback} from "react";
-import {Controller} from "react-hook-form";
-import {AdditionalProps, Button, ButtonIconbag, GeneralIcon, Input, useTrans} from "zographia";
-import {UserSocialTypeSelect} from "/client/component/atom/user-social-type-select";
+import {ReactElement} from "react";
+import {AdditionalProps, Button, ButtonIconbag, GeneralIcon, useTrans} from "zographia";
 import {create} from "/client/component/create";
-import {UserSocialType, UserWithDetail} from "/server/internal/skeleton";
+import {UserWithDetail} from "/server/internal/skeleton";
 import {SOCIAL_ROW_COUNT, useChangeMySocials} from "./change-my-socials-form-hook";
+import {ChangeMySocialsFormRow} from "./change-my-socials-form-row";
 
 
 export const ChangeMySocialsForm = create(
@@ -23,25 +22,12 @@ export const ChangeMySocialsForm = create(
     const {trans} = useTrans("changeMySocialsForm");
 
     const {form, handleSubmit} = useChangeMySocials(me);
-    const {control, register, setValue} = form;
-
-    const handleBlur = useCallback(function (event: FocusEvent<HTMLInputElement>): void {
-      const type = guessUserSocialType(event.target.value);
-      if (type !== null) {
-        setValue(event.target.name.replace(/\.url$/, ".type") as `socials.${number}.type`, type);
-      }
-    }, [setValue]);
 
     return (
       <form styleName="root" {...rest}>
         <div styleName="table">
           {Array.from({length: SOCIAL_ROW_COUNT}, (dummy, index) => (
-            <div styleName="row" key={index}>
-              <Controller name={`socials.${index}.type`} control={control} render={({field}) => (
-                <UserSocialTypeSelect styleName="type" type={field.value} onSet={field.onChange}/>
-              )}/>
-              <Input styleName="url" placeholder={trans("placeholder.url")} {...register(`socials.${index}.url`, {onBlur: handleBlur})}/>
-            </div>
+            <ChangeMySocialsFormRow key={index} form={form} index={index}/>
           ))}
         </div>
         <div>
@@ -55,28 +41,3 @@ export const ChangeMySocialsForm = create(
 
   }
 );
-
-
-function guessUserSocialType(url: string): UserSocialType | null {
-  try {
-    const normalizedUrl = (/^https?:\/\//.test(url)) ? url : `https://${url}`;
-    const host = new URL(normalizedUrl).hostname.replace(/^www\./, "");
-    if (host === "x.com" || host === "twitter.com") {
-      return "x";
-    } else if (host === "bsky.app") {
-      return "bluesky";
-    } else if (host.includes("misskey")) {
-      return "misskey";
-    } else if (host === "note.com") {
-      return "note";
-    } else if (host === "migdal.jp") {
-      return "migdal";
-    } else if (host === "discord.gg" || host === "discord.com" || host === "discordapp.com") {
-      return "discord";
-    } else {
-      return null;
-    }
-  } catch (error) {
-    return null;
-  }
-}
